fix: cleanup code and ensure qt5 in the cmake file

+++ src/waterfall_raster_data.cpp
@@ -1,12 +1,11 @@
 #include "waterfall_raster_data.hpp"
 
 #include <QtGlobal>
-#include <vector>
 #include <iostream>
+#include <vector>
 
 WaterfallRasterData::WaterfallRasterData(int rows, int cols, int bin_width, double init_value)
-    : m_maxRows(rows), m_cols(cols), bin_width(bin_width), m_currentIndex(0), init_value(init_value)
-{
+    : m_maxRows(rows), m_cols(cols), bin_width(bin_width), m_currentIndex(0), init_value(init_value) {
     m_data.resize(m_maxRows * m_cols, init_value);
 
     setInterval(Qt::XAxis, QwtInterval(0, m_cols));
@@ -20,7 +19,7 @@ WaterfallRasterData::~WaterfallRasterData() {
 
 void WaterfallRasterData::addRow(QVector<double> newRow) {
     int start_index = m_currentIndex * m_cols;
-    for (int i = 0; i < std::min(m_cols, newRow.size()); ++i) {
+    for (int i = 0; i < std::min(m_cols, static_cast<int>(newRow.size())); ++i) {
         m_data[start_index + i] = newRow.at(i);
     }
 
@@ -33,7 +32,7 @@ void WaterfallRasterData::addRow(std::map<uint64_t, float> newRow) {
 
     QVector<double> newRowVector(m_cols, init_value);
 
-    for (auto& pair: newRow) {
+    for (auto& pair : newRow) {
         newRowVector[(pair.first - start_freq) / bin_width] = pair.second;
     }
